test: 방 생성/삭제 통합 테스트 분리 #11
